requirements.txt
asgiref==3.10.0
certifi==2025.10.5
chardet==4.0.0
dj-database-url==3.0.1
Django==5.2.7
django-decouple==2.1
django-environ==0.12.0
django-parler==2.3
django-redis==6.0.0
django-rosetta==0.9.9
gunicorn==23.0.0
idna==2.10
packaging==25.0
pillow==12.0.0
polib==1.2.0
psycopg2==2.9.11
redis==7.0.1
requests==2.25.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==1.26.20
